% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoRec.R
\docType{package}
\name{PhenoRec-package}
\alias{PhenoRec}
\alias{PhenoRec-package}
\title{PhenoRec: phenotype-preserving integration of single-cell data}
\description{
PhenoRec provides functions for recovering phenotype-associated variation
that may be removed during conventional single-cell data integration.
The current implementation supports Gaussian measurements (for example,
normalized RNA expression) and Bernoulli measurements (for example, binarized
chromatin accessibility).
}
\author{
\strong{Maintainer}: Chao Deng \email{deng_chao@sjtu.edu.cn}

}
\keyword{internal}
